Add conformance to OnboardingViewControllerProtocol
